Fixed various stuff
Fixed propagation of pause and resume for linux client. These events is
propagated on window focus gained/loss. Also fixed so that gamepad
events is discarded if window is not focused.

Furthermore added means to access the aabb constraining the view port of
a scene, this is to ease positioning of ui components.

And finally updated so that it is possible to render texts with an
offset. We have some refactoring to do here since it's not good design
that Text is both a rendering component and a renderable.